Moved test_output_current_state into core/data tests
